Moved redact paths to another file for unit testing issue
